+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Upcoming
+- Fix locale methods
+- Change API of locale methods
+- Add locale update method
+
 ## v0.3.0 - 2015-10-16
 - Add methods to get published entities